fixed review

+++ app/controllers/foods_controller.rb
@@ -1,6 +1,6 @@
 class FoodsController < ApplicationController
   def index
-    @foods = Food.all
+    @foods = current_user.foods
   end
 
   def new
@@ -20,12 +20,16 @@ class FoodsController < ApplicationController
   def destroy
     food = Food.find(params[:id])
 
-    if food.destroy
-      flash[:notice] = 'Food was successfully deleted.'
+    if food.user == current_user
+      if food.destroy
+        flash[:notice] = 'Food was successfully deleted.'
+      else
+        flash[:alert] = 'Food deleting Failed. Please try again.'
+      end
+      redirect_back(fallback_location: root_path)
     else
-      flash[:alert] = 'Food deleting Failed. Please try again.'
+      flash[:alert] = 'You do not have access to delete the Food belongs to other Users.'
     end
-    redirect_back(fallback_location: root_path)
   end
 
   def food_params
